Added confirmation message from delete action

+++ UsedCarShowroomApp/src/main/webapp/home.jsp
@@ -49,7 +49,7 @@
 						<td>
 							<a href="edit?carId=${car.carId}">Edit</a>
 							&nbsp;&nbsp;&nbsp;&nbsp;
-							<a href="delete?carId=${car.carId}">Delete</a>
+							<a href="delete?carId=${car.carId}" onclick="return confirm('Are you sure you want to delete?')">Delete</a>
 						</td>
 						<% } %>
 					</tr>
